+++ extensions/vscode/src/utils/get-template-setting.ts
