+++ extensions/vscode/src/utils/index.ts
@@ -3,7 +3,7 @@ export * from "./get-bloc-type";
 export * from "./get-pubspec";
 export * from "./get-pubspec-path";
 export * from "./get-selected-text";
-export * from "./get-template-settings";
+export * from "./get-template-setting";
 export * from "./has-dependency";
 export * from "./update-pubspec-dependency";
 export * from "./wrap-with";
